+++ src/main/baseball/Validation.java
@@ -3,10 +3,10 @@ package baseball;
 import java.util.List;
 
 public class Validation {
-    private BallsMaker ballsMaker;
+    private Balls balls;
 
-    public Validation(BallsMaker ballsMaker) {
-        this.ballsMaker = ballsMaker;
+    public Validation(Balls balls) {
+        this.balls = balls;
     }
 
     public void validateInput(String input) {
@@ -14,7 +14,7 @@ public class Validation {
             throw new IllegalArgumentException("입력은 " + Random.SIZE + "자리여야 합니다.");
         }
 
-        List<Integer> playerNumbers = ballsMaker.splitStr(input);
+        List<Integer> playerNumbers = balls.splitStr(input);
 
         if (playerNumbers.stream().anyMatch(num -> !Ball.validNo(num)) || Ball.hasDuplicate(playerNumbers)) {
             throw new IllegalArgumentException("유효하지 않은 입력입니다. 다시 입력해주세요.");
